+++ src/DataTable/DataTable.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component, memo } from 'react';
 import PropTypes from 'prop-types';
 import styles from './DataTable.scss';
 import classNames from 'classnames';
@@ -10,6 +10,7 @@ import { Animator } from 'wix-animations';
 import Tooltip from '../Tooltip/Tooltip';
 import InfoIcon from '../common/InfoIcon';
 import { VariableSizeList as List } from 'react-window';
+import { virtualRowsAreEqual } from './DataTable.utils';
 
 export const DataTableHeader = props => {
   const { dataHook } = props;
@@ -29,7 +30,7 @@ DataTableHeader.propTypes = {
 class DataTable extends React.Component {
   constructor(props) {
     super(props);
-    let state = { selectedRows: {} };
+    let state = { selectedRows: new Map() };
     if (props.infiniteScroll) {
       state = { ...state, ...this.createInitialScrollingState(props) };
     }
@@ -142,7 +143,7 @@ class DataTable extends React.Component {
   onRowClick = (rowData, rowNum) => {
     const { onRowClick, rowDetails } = this.props;
     onRowClick && onRowClick(rowData, rowNum);
-    rowDetails && this.toggleRowDetails(rowNum);
+    rowDetails && this.toggleRowDetails(rowData);
   };
 
   renderRow = (rowData, rowNum, style) => {
@@ -228,7 +229,7 @@ class DataTable extends React.Component {
     ];
 
     if (rowDetails) {
-      const showDetails = !!this.state.selectedRows[rowNum];
+      const showDetails = !!this.state.selectedRows.get(rowData);
 
       rowsToRender.push(
         <tr
@@ -294,24 +295,35 @@ class DataTable extends React.Component {
   };
 
   toggleRowDetails = selectedRow => {
-    let selectedRows = { [selectedRow]: !this.state.selectedRows[selectedRow] };
-    if (this.props.allowMultiDetailsExpansion && !this.props.virtualized) {
-      selectedRows = Object.assign({}, this.state.selectedRows, {
-        [selectedRow]: !this.state.selectedRows[selectedRow],
-      });
-    }
-    this.setState({ selectedRows });
+    const { selectedRows } = this.state;
+
+    const allowMultipleRowDetails =
+      this.props.allowMultiDetailsExpansion && !this.props.virtualized;
+
+    const newSelectedRows = new Map([
+      ...(allowMultipleRowDetails ? [...selectedRows] : []),
+      [selectedRow, !selectedRows.get(selectedRow)],
+    ]);
+
+    this.setState({ selectedRows: newSelectedRows });
   };
 
-  renderVirtualizedRow = ({ index, style }) =>
-    this.renderRow(this.props.data[index], index, style)[0];
+  renderVirtualizedRow = ({ data, index, style }) =>
+    this.renderRow(data[index], index, style)[0];
+
+  renderVirtualizedMemoizedRow = virtualListProps =>
+    React.createElement(this.memoizedRow, {
+      ...this.props,
+      ...virtualListProps,
+    });
+
+  memoizedRow = memo(this.renderVirtualizedRow, virtualRowsAreEqual);
 
   getVirtualRowHeight = () => this.props.virtualizedLineHeight;
 
-  renderVirtualizedTableElementWithRefForward = () =>
-    React.forwardRef((props, ref) =>
-      this.renderVirtualizedTableElement({ ...props, ref }),
-    );
+  virtualizedTableElementWithRefForward = React.forwardRef((props, ref) =>
+    this.renderVirtualizedTableElement({ ...props, ref }),
+  );
 
   renderVirtualizedTableElement = ({ children, ...rest }) => {
     return (
@@ -330,12 +342,13 @@ class DataTable extends React.Component {
           className={classNames(this.style.table, this.style.virtualized)}
           height={virtualizedTableHeight}
           itemCount={data.length}
+          itemData={data}
           width={'100%'}
           itemSize={this.getVirtualRowHeight}
-          outerElementType={this.renderVirtualizedTableElementWithRefForward()}
+          outerElementType={this.virtualizedTableElementWithRefForward}
           innerElementType={'tbody'}
         >
-          {this.renderVirtualizedRow}
+          {this.renderVirtualizedMemoizedRow}
         </List>
       </div>
     );