+++ src/SidebarDivider/test/SidebarDivider.visual.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 
 import SidebarDivider from '../SidebarDivider';
-import Sidebar from '../../Sidebar';
 import Box from '../../Box';
+import { SidebarContext } from '../../Sidebar/SidebarAPI';
 
 const skins = ['dark', 'light'];
 
@@ -32,10 +32,10 @@ tests.forEach(({ describe, its }) =>
         <Box backgroundColor="D70">
           {skins.map(skin => (
             <Box direction="vertical" marginBottom={5} marginRight={5}>
-              <div style={{ width: '222px' }}>
-                <Sidebar skin={skin}>
+              <div style={{ width: '228px' }}>
+                <SidebarContext.Provider value={{ getSkin: () => skin }}>
                   <SidebarDivider {...props} />
-                </Sidebar>
+                </SidebarContext.Provider>
               </div>
             </Box>
           ))}