+++ src/SortableList/SortableList.spec.js
@@ -338,6 +338,78 @@ describe('SortableList', () => {
     expect(props.onDragStart).not.toBeCalled();
   });
 
+  it('should insert at end of the list', () => {
+    const dataHook = 'sortable-list';
+    const items = [
+      { id: '1', text: 'item 1' },
+      { id: '2', text: 'item 2' },
+      { id: '3', text: 'item 3' },
+    ];
+    const onDrop = jest.fn();
+    const renderItem = ({ item }) => <div>{item.text}</div>; // eslint-disable-line react/prop-types
+
+    const wrapper = ReactTestUtils.renderIntoDocument(
+      <DragDropContextProvider backend={TestBackend}>
+        <SortableList
+          contentClassName="cl"
+          dataHook={dataHook}
+          containerId="sortable-list"
+          items={items}
+          insertPosition="end"
+          renderItem={renderItem}
+          onDrop={onDrop}
+        />
+      </DragDropContextProvider>,
+    );
+    const driver = sortableListTestkitFactory({ wrapper, dataHook });
+
+    driver.reorder({ removedId: '1', addedId: '2' });
+
+    expect(onDrop).toBeCalledWith({
+      addedIndex: 2,
+      addedToContainerId: 'sortable-list',
+      payload: { id: '1', text: 'item 1' },
+      removedFromContainerId: 'sortable-list',
+      removedIndex: 0,
+    });
+  });
+
+  it('should insert at start of the list', () => {
+    const dataHook = 'sortable-list';
+    const items = [
+      { id: '1', text: 'item 1' },
+      { id: '2', text: 'item 2' },
+      { id: '3', text: 'item 3' },
+    ];
+    const onDrop = jest.fn();
+    const renderItem = ({ item }) => <div>{item.text}</div>; // eslint-disable-line react/prop-types
+
+    const wrapper = ReactTestUtils.renderIntoDocument(
+      <DragDropContextProvider backend={TestBackend}>
+        <SortableList
+          contentClassName="cl"
+          dataHook={dataHook}
+          containerId="sortable-list"
+          items={items}
+          insertPosition="start"
+          renderItem={renderItem}
+          onDrop={onDrop}
+        />
+      </DragDropContextProvider>,
+    );
+    const driver = sortableListTestkitFactory({ wrapper, dataHook });
+
+    driver.reorder({ removedId: '1', addedId: '2' });
+
+    expect(onDrop).toBeCalledWith({
+      addedIndex: 0,
+      addedToContainerId: 'sortable-list',
+      payload: { id: '1', text: 'item 1' },
+      removedFromContainerId: 'sortable-list',
+      removedIndex: 0,
+    });
+  });
+
   it('should contain prop to set custom  class (`isListInDragState`) while dragging', () => {
     const renderItem = ({ item, isListInDragState }) => (
       <div
@@ -458,6 +530,57 @@ describe('Enzyme: SortableList', () => {
       removedFromContainerId: 'sortable-list',
       removedIndex: 0,
     });
+  });
+
+  it('should not call onDrop when drag between columns in case of droppable false', () => {
+    const dataHook = 'sortable-list';
+    const items = [{ id: '1', text: 'item 1' }, { id: '2', text: 'item 2' }];
+    const items2 = [
+      { id: '11', text: 'item 11' },
+      { id: '21', text: 'item 21' },
+    ];
+    const onDrop = jest.fn();
+    const renderItem = ({ item }) => <div key={item.id}>{item.text}</div>; // eslint-disable-line react/prop-types
+
+    class MyComponent extends React.Component {
+      render() {
+        return (
+          <div>
+            <SortableList
+              contentClassName="cl"
+              dataHook={dataHook}
+              containerId="sortable-list-1"
+              groupName="group1"
+              items={items}
+              renderItem={renderItem}
+              onDrop={onDrop}
+            />
+            <SortableList
+              canDrop={() => false}
+              contentClassName="cl"
+              dataHook={dataHook}
+              containerId="sortable-list-2"
+              groupName="group1"
+              items={items2}
+              droppable={false}
+              renderItem={renderItem}
+              onDrop={onDrop}
+            />
+          </div>
+        );
+      }
+    }
+
+    const wrapper = mount(
+      <DragDropContextProvider backend={TestBackend}>
+        <MyComponent />
+      </DragDropContextProvider>,
+    );
+    const driver = enzymeSortableListTestkitFactory({ wrapper, dataHook });
+
+    driver.reorder({ removedId: '1', addedId: '21' });
+
+    expect(onDrop).not.toHaveBeenCalled();
   });
 
   it('should call onDrop when drag between columns', () => {