+++ src/Swatches/Swatches.js
@@ -26,6 +26,8 @@ const Swatches = props => {
     addButtonMessage,
     addButtonIconSize,
     onAdd,
+    onChange,
+    onCancel,
     columns,
     gap,
   } = props;
@@ -48,6 +50,8 @@ const Swatches = props => {
           tooltip={addButtonMessage}
           iconSize={addButtonIconSize}
           onAdd={onAdd}
+          onChange={onChange}
+          onCancel={onCancel}
         />
       )}
 
@@ -104,6 +108,12 @@ Swatches.propTypes = {
 
   /** Callback function when user clicks on Add button and selects a color to be added. Returns color HEX string representation. */
   onAdd: func,
+
+  /** Callback function when user changes color in Color Picker. Returns color HEX string representation. */
+  onChange: func,
+
+  /** Callback function when user closes Color Picker without picking color. */
+  onCancel: func,
 
   /** If true shows add button which triggers colors picker*/
   showAddButton: bool,