+++ testkit/enzyme.d.ts
@@ -122,6 +122,7 @@ declare namespace EnzymeTestkit {
   export const toggleSwitchTestkitFactory: any;
   export const sidebarSectionItemTestkitFactory: any;
   export const sidebarDividerTestkitFactory: any;
+  export const sidebarBackButtonTestkitFactory: any;
   export const sidebarHeaderTestkitFactory: any;
   export const modalPreviewLayoutTestkitFactory: any;
   export const statisticsWidgetTestkitFactory: any;
@@ -133,6 +134,9 @@ declare namespace EnzymeTestkit {
   export const autoCompleteWithLabelTestkitFactory: any;
   export const dividerTestkitFactory: any;
   export const labelledElementTestkitFactory: any;
+  export const previewWidgetTestkitFactory: any;
+  export const modalMobileLayoutTestkitFactory: any;
+  export const mediaOverlayTestkitFactory: any;
   export const sideMenuDrillTestkitFactory: any;
   export const headerTestkitFactory: any;
   export const sideMenuTestkitFactory: any;