+++ testkit/index.d.ts
@@ -118,6 +118,7 @@ declare namespace VanillaTestkit {
   export const toggleSwitchTestkitFactory: any;
   export const sidebarSectionItemTestkitFactory: any;
   export const sidebarDividerTestkitFactory: any;
+  export const sidebarBackButtonTestkitFactory: any;
   export const sidebarHeaderTestkitFactory: any;
   export const modalPreviewLayoutTestkitFactory: any;
   export const statisticsWidgetTestkitFactory: any;
@@ -129,6 +130,9 @@ declare namespace VanillaTestkit {
   export const autoCompleteWithLabelTestkitFactory: any;
   export const dividerTestkitFactory: any;
   export const labelledElementTestkitFactory: any;
+  export const previewWidgetTestkitFactory: any;
+  export const modalMobileLayoutTestkitFactory: any;
+  export const mediaOverlayTestkitFactory: any;
   export const sideMenuDrillTestkitFactory: any;
   export const headerTestkitFactory: any;
   export const sideMenuTestkitFactory: any;